Mejor ordenado y nuevos avances
